refactor: merge ContractFunction and AsyncContractFunction code
