+++ web3/types.py
@@ -38,11 +38,12 @@ from web3._utils.compat import (
 if TYPE_CHECKING:
     from web3.contract.async_contract import (  # noqa: F401
         AsyncContractEvent,
-        AsyncContractFunction,
+    )
+    from web3.contract.base_contract import (
+        BaseContractFunction,
     )
     from web3.contract.contract import (  # noqa: F401
         ContractEvent,
-        ContractFunction,
     )
     from web3.main import (  # noqa: F401
         AsyncWeb3,
@@ -581,7 +582,7 @@ class GethWallet(TypedDict):
 
 # Contract types
 
-TContractFn = TypeVar("TContractFn", "ContractFunction", "AsyncContractFunction")
+TContractFn = TypeVar("TContractFn", bound="BaseContractFunction")
 TContractEvent = TypeVar("TContractEvent", "ContractEvent", "AsyncContractEvent")
